client2: fix widget destroy hook
Previously, `Widget.destroy` only got called when a node got orphaned by a
node clearing operation, and collected by
`LonaRenderingEngine._clean_node_cache`. It did not run, when a single node got
removed from the client.

This patch moves the destroy and clean-up code into a helper method of the
rendering engine, and adds a call to it to the node removal code, so
`Widget.destroy` gets called every time a node gets removed from the client.

Signed-off-by: Florian Scherf <mail@florianscherf.de>