client: fix widget deconstruct hook
Previously, `Widget.deconstruct` only got called when a node got orphaned by a
node clearing operation, and collected by `_clean_node_cache`. It did not run,
when a singe node got removed from the client.

This patch moves the deconstruct and clean-up code into a helper method, and
adds a call to it to the node removal code, so `Widget.deconstruct` gets called
every time a node gets removed from the client.

Signed-off-by: Florian Scherf <mail@florianscherf.de>