rendering: add support for namespaces
Previously, the rendering only was able to render nodes from the xhtml
namespace. This patch adds a namespace property to all nodes, and adds
namespace support to both client and client2.
The default namespace is '' which is replaced with the xhtml namespace on both
clients.

This change will make SVG rendering possible in the future.

Signed-off-by: Florian Scherf <mail@florianscherf.de>