html: parsing: fix parsing of single nodes in HTML-strings
Previously, when parsing HTML-strings, that contained only one node, the
resulting node was no root node, so it had a parent node set that was out of
scope.

As an implementation detail, the parsing code sets an arbitrary root node while
parsing, and in this particular case, this root node was not unset correctly.

This patch updates the parsing code to only return root nodes.

Signed-off-by: Florian Scherf <mail@florianscherf.de>