html: parsing: fork `html.parser.HTMLParser.parse_starttag`
Lona uses `html.parser.HTMLParser` to convert HTML-strings to Lona nodes. This
implementation converts all attribute names, of HTML tags, to lower case in
`html.parser.HTMLParser.parse_starttag`, which is fine for HTML, but becomes a
problem when parsing SVGs.

To fix this in the future, this patch forks the method from the standard
library, without changing it.
This patch also adds some NOQAs to make the linter happy.

Signed-off-by: Florian Scherf <mail@florianscherf.de>