html: node: Remove unnecessary index() from NodeList.append()
Calling index() to get the previously appended Node's position is
unnecessary here, since it will by default be in the list's last
position.

This patch also improves the overall performance of append() calls,
since index() currently relies on eq(), which is implemented rather
expensively.

Avoiding index() takes a

        tr = Tr()
        for i in range(100):
            tr.append(Td(i))

from 10'000 calls of serialize() to 100.