client2: rendering: fix caching problem when slicing child nodes
Previously, when running a view that did some kind of node slicing, or
resending nodes in general, like this:

    node = Div(
        Div('div 1'),
        Div('div 2'),
        Div('div 3'),
    )

    self.show(node)

    node.nodes = node.nodes[1:-1]

    self.show(node)

the client would crash raising a runtime error, saying:

    lona.view_runtime client error raised: node with id $NODE is already cached

This happened because the rendering engine did not clean its node cache after
a RESET operation was handled. That means the node gets cleared, and the
rendering engine attempts to re-enter some of the previous child nodes, but
because their node ids are still present in the node cache, a safety mechanism
gets triggered, complaining that a node tries to be rendered twice, which is
wrong in this case.

This patch changes the handling of RESET operations, to clean the node cache
after a node got cleared, and also adds a test that tests node slicing.

Signed-off-by: Florian Scherf <mail@florianscherf.de>