html: AbstractNode: improve node compare performance
Previously, `AbstractNode.__eq__` used `AbstractNode._serialize`, which
converts a node to JSON, to compare two nodes with each other. This worked
fine but had a significant overhead.

This patch updates `AbstractNode.__eq__` to compare all node attributes with
each other, trying to find differences as early, and therefor as cheap, as
possible.

Signed-off-by: Florian Scherf <mail@florianscherf.de>