client: add support for reconnect without creating a window
Previously, when implementing auto-reconnect, like in the default template of
lona-picocss, the client would reopen the websocket connection and in the case
of success reload the tab. This reload is crucial to ensure a connect and a
reconnect result in the same user experience, but has the side effect of
accessing the same view twice, which creates problems when debugging or reading
the server logs.

This patch adds an option `create_window`, which by default is set to `true`,
to `LonaContext.reconnect()`. When set to `false` the client will only
reconnect the WebSocket, but won't create a window and therefor access a view
on the server.

Signed-off-by: Florian Scherf <mail@florianscherf.de>