client1 + client2: Fix `onsubmit` event handler to allow submitting
According to the documentation
(https://lona-web.org/1.x/api-reference/views.html#post ) it should be
possible to submit a `<form>` using `POST` to an interactive lona view.
Without this change this fails with a
`Uncaught ReferenceError: lona_window is not defined` in the browser.

This change fixes the missing reference and allows to use the `POST` method
to submit data to the view.

(Also add a test to make sure a POSTed `<form>` still works in the
future.)

Signed-off-by: Chris Fiege <cfi@pengutronix.de>