client2: rendering: fix call stack limit problems in node cache
Previously, the rendering of client 2 would crash when too many nodes were
removed at the same time. This happened because the rendering engine method
for removing a node, called the method for cleaning the node cache, and
vice versa.
In theory that is correct and works, but if the amount of nodes that should be
removed is big enough, the JavaScript call stack size will exceed the browsers
limit.

This patch removes the dependency between the two methods and adds more careful
cache cleaning calls.

Signed-off-by: Florian Scherf <mail@florianscherf.de>