client: fix problems with the browser history
The browser history of the client never *really* worked. When you were
redirected to a Lona app from an external site, you couldn't go back by
clicking on the back button. When you were three links into an Lona app and
refreshed the tab, the history before the refresh was unreachable.

Previously, both client1 and client2 created a new history entry, using
`history.pushState`, when a new view started, regardless the reason of the view
starting. So initial setup, refreshes, and reconnects also created history
items.
This is not correct! `history.pushState` should only be called when
you want to *extend* the browser history, beyond the normal history behaviour.

This patch limits all calls to `history.pushState` to cases of actual in-app
navigation, and adds a test to check the now corrected behaviour.

Signed-off-by: Florian Scherf <mail@florianscherf.de>