tests: playwright: bump from 1.44.0 to 1.44.0
Somewhere between Playwright 1.40 and 1.44, `Page.wait_for_url()` changed
behavior. Now, only URLs, starting with wildcards, seem to work for
our usecase.

I am not sure why that was changed, or why this is a problem in our test, but
the change in this patch is in line with the documentation of
`Page.wait_for_url()`, so it should be fine.

Signed-off-by: Florian Scherf <mail@florianscherf.de>